pyproject.toml
[tool.poetry]
name = "rest-csv-reader"
version = "0.1.0"
description = ""
authors = ["Nanashi <alksndr.zln@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
django = "3.2.18"
django-rest-framework = "^0.1.0"
environs = "^9.5.0"
drf-spectacular = "^0.26.3"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
